refactor: Constants 클래스 패키지 변경
