+++ tomcat/src/main/java/org/apache/Constants.java
@@ -1,4 +1,4 @@
-package common;
+package org.apache;
 
 public class Constants {
