+++ tomcat/src/test/java/nextstep/jwp/controller/HomeControllerTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
 
-import static common.Constants.CRLF;
+import static org.apache.Constants.CRLF;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
