+++ tomcat/src/test/java/nextstep/jwp/controller/LoginControllerTest.java
@@ -12,7 +12,7 @@ import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
 
-import static common.Constants.CRLF;
+import static org.apache.Constants.CRLF;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
